refactor: remove IVideoCommand interface and related adapter, update use cases to use IVideoAggregateRepository
