+++ src/infrastructure/adapters/video/VideoMetadataExtractorAdapter.ts
